Macro for micro-card. Refactor for kode-gjenbruk
