+++ src/types/macro-props/product-card-mini.ts
@@ -1,9 +1,9 @@
 import { MacroPropsCommon, MacroType } from './_macros-common';
-import { ProductTarget } from '../component-props/parts/product-card';
+import { ProductCardMiniProps } from '../component-props/parts/product-card';
 
 export interface MacroProductCardMiniProps extends MacroPropsCommon {
     name: MacroType.ProductCardMini;
     config: {
-        product_card_mini: ProductTarget;
+        product_card_mini: ProductCardMiniProps['config'];
     };
 }
